Enable automatic certificates for production.
The certificate is renewed using the `montagu renew-certificate`
command. A system timer and one-shot service are used to do this
periodically. While the timer is set to run daily, in practice certbot
checks the existing certificate, and only renews it if the certificate
is due to expire in the next 30 days..