Update renew-certificate periodic task.
The interface for the montagu-deploy tool has changed a bit.

The working directory should now be the root of the montagu-config
checkout. We do not need to save the name of the deployment anymore as
it is captured by the `.montagu_identity` file already.

All of this will get removed once we migrate to acme-buddy, when we
won't need an external timer to run anymore.